backend/requirements.txt
fastapi
numpy
Pillow
python-multipart
uvicorn
